pyproject.toml
[tool.poetry]
name = "pyuno"
version = "0.1.0"
description = ""
authors = ["Kleomedes Dev Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
cosmpy = "==0.6.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "==3.2.0"
check-manifest = "*"
tox = "==3.25.1"
black = "==23.1"
darglint = "==1.8.1"
grpcio-tools = "==1.47.0"
ruff = "==0.0.259"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "==8.4"
pydoc-markdown = "==4.6.3"


[tool.poetry.group.test.dependencies]
coverage = "*"
pytest = "*"
pytest-rerunfailures = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "pyuno/protos"
]

# Same as Black.
line-length = 100

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
